pyproject.toml
[tool.poetry]
name = "myapp"
version = "0.1.0"
description = "Sample python for Tensorflow on MacOS"
authors = ["phdcybersec <82591009+phdcybersec@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
tensorflow-metal = "~0.6.0"
tensorflow-macos = "~2.10.0"
grpcio = ">=1.37.0,<2.0"
h5py = ">=3.6.0,<3.7"
numpy = ">=1.23.2,<1.23.3"
protobuf = ">=3.19.1,<3.20"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
autoflake = "^2.1.1"

[tool.poetry.scripts]
myapp = "myapp:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
